Create flutter-ci.yml
This ci does whatever the previous ci (cleaner code tho) did but it additionally uploads the apk for each PR to github artifacts which can be donwloaded by the qa team to make their testing easier. So basically the tester from qa team will not have to use an android emulator or install flutter to test PRs on this repo anymore. Just download the artifact from below each PR :)